move gather_ndarray to util submodule
